switch to real service in e2e test
